Update run.sh
Add If Statement

+++ run.sh
@@ -253,10 +253,13 @@ function install_dependencies() {
 }
 
 function start_ollama(){
-	wget "https://github.com/ollama/ollama/releases/download/v0.1.20/ollama-linux-amd64"
+	if [ ! -f $SCRATCH/ollama]; then 
+ 	wget "https://github.com/ollama/ollama/releases/download/v0.1.20/ollama-linux-amd64"
 	chmod 755 ollama-linux-amd64
 	mv ollama-linux-amd64 $SCRATCH/ollama
-	nohup $SCRATCH/ollama serve &
+	
+ 	fi 
+  	nohup $SCRATCH/ollama serve &
 }
 
 #Parameters
